+++ packages/sui-studio/bin/sui-studio-dev.js
@@ -48,8 +48,10 @@ const legacyTestPath = path.join(PWD, 'test', category, component)
 
 const testPath = fs.existsSync(legacyTestPath) ? legacyTestPath : path.join(componentPath, 'test')
 
+const {cache, ...others} = config
+
 const studioDevConfig = {
-  ...config,
+  ...others,
   context: path.join(__dirname, '..', 'workbench', 'src'),
   plugins: [...config.plugins, new webpack.DefinePlugin({__COMPONENT_ID__: JSON.stringify(componentID)})],
   resolve: {