+++ packages/sui-test/bin/mocha/register.js
@@ -9,7 +9,7 @@ if (esmOverride) {
 }
 
 const libDir = /lib/
-const paths = [/@babel\/runtime/, /@s-ui/, /mocks/, /src/, /test/, libDir, ...regexToAdd]
+const paths = [/@babel\/runtime/, /@s-ui/, /@adv-ui/, /mocks/, /src/, /test/, libDir, ...regexToAdd]
 const only = useLibDir ? paths : paths.filter(path => path !== libDir)
 const swcConfig = getSWCConfig({isTypeScript: true, compileToCJS: true})