+++ packages/sui-studio/CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+# 11.38.0 (2024-01-05)
+
+
+### Features
+
+* **packages/sui-studio:** disable filesystem cache in dev mode ([200c5f6](https://github.com/SUI-Components/sui/commit/200c5f610c919bd60f6493eb57182ff2791feaa6))
+
+
+
+# 11.37.0 (2023-12-19)
+
+
+
 # 11.36.0 (2023-08-17)